[SCRUM-18] chore: 피드 도메인 분리
