+++ src/main/java/com/sprint/otboo/feed/service/LikeService.java
@@ -1,0 +1,7 @@
+package com.sprint.otboo.feed.service;
+
+import java.util.UUID;
+
+public interface LikeService {
+    void addLike(UUID feedId, UUID userId);
+}
