+++ src/test/java/com/sprint/otboo/feed/controller/FeedControllerTest.java
@@ -22,7 +22,6 @@ import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
@@ -30,7 +29,6 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import static org.mockito.BDDMockito.given;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.willDoNothing;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
@@ -160,29 +158,6 @@ class FeedControllerTest {
                 )
                 .andExpect(status().isBadRequest())
                 .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON));
-        }
-    }
-
-    @Nested
-    @DisplayName("피드 좋아요 등록 테스트")
-    class FeedLikeCreateTests {
-
-        @Test
-        @DisplayName("좋아요를_등록하면_204를_반환한다")
-        void 좋아요를_등록하면_204를_반환한다() throws Exception {
-            // Given
-            UUID feedId = UUID.randomUUID();
-            UUID userId = UUID.randomUUID();
-
-            willDoNothing().given(feedService).addLike(feedId, userId);
-
-            // When & Then
-            mockMvc.perform(post("/api/feeds/{feedId}/like", feedId)
-                .with(csrf())
-                .with(user("tester").roles("USER"))
-                    .header("X-USER-ID", userId.toString())
-                    .contentType(MediaType.APPLICATION_JSON))
-                .andExpect(status().isNoContent());
         }
     }
 }
